+++ web/src/pages/apd/ApdPageRoutes.js
@@ -11,7 +11,7 @@ import ApdHeader from '../../layout/header/ApdHeader';
 import Activities from '../../containers/activity/All';
 import EntryPage from '../../containers/activity/EntryPage';
 import AssurancesAndCompliance from './assurances-and-compliance/AssurancesAndCompliance';
-import Export from '../../containers/ApdExport';
+import Export from './export/ApdExport';
 import ApdOverview from './apd-overview/ApdOverview';
 import ExecutiveSummary from './executive-summary/ExecutiveSummary';
 import PreviousActivities from './previous-activities/PreviousActivities';
