+++ web/src/pages/apd/ApdReadOnly.js
@@ -20,7 +20,7 @@ import ScheduleSummary from './schedule-summary/ScheduleSummaryReadOnly';
 import ProposedBudget from './proposed-budget/ProposedBudgetReadOnly';
 import AssuranceAndCompliance from './assurances-and-compliance/AssurancesAndComplianceReadOnly';
 import ExecutiveSummary from './executive-summary/ExecutiveSummaryReadOnly';
-import ExportInstructions from '../../containers/viewOnly/Export';
+import ExportInstructions from './export/ExportReadOnly';
 import Loading from '../../components/Loading';
 
 const ApdViewOnly = ({
